backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
click==8.3.1
fastapi==0.122.0
h11==0.16.0
httptools==0.7.1
httpx==0.28.1
idna==3.11
pdfplumber==0.11.9
pydantic==2.12.5
pydantic_core==2.41.5
pymupdf==1.25.5
python-dotenv==1.2.2
PyYAML==6.0.3
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
anthropic>=0.94.0
weasyprint==68.1
